+++ src/Asset/LazyAsset.php
@@ -9,12 +9,12 @@
  * file that was distributed with this source code.
  */
 
-namespace Puli\Extension\Assetic\Asset;
+namespace Puli\AsseticExtension\Asset;
 
 use Assetic\Asset\AssetInterface;
 use Assetic\Filter\FilterCollection;
 use Assetic\Filter\FilterInterface;
-use Puli\Extension\Assetic\Factory\PuliAssetFactory;
+use Puli\AsseticExtension\Factory\PuliAssetFactory;
 use RuntimeException;
 
 /**