+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 * 1.0.0-alpha2 (@release_date@)
 
  * renamed `PuliAssetInterface` to `PuliAsset`
+ * renamed `Deferred*` classes to `Lazy*`
  
 * 1.0.0-alpha1 (2014-12-03)