+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 Changelog
 =========
 
-* 1.0.0-alpha2 (@release_date@)
+* 1.0.0-next (@release_date@)
 
  * renamed `PuliAssetInterface` to `PuliAsset`
  * renamed `Deferred*` classes to `Lazy*`
+ * moved all code to `Puli\AsseticExtension` namespace
  
 * 1.0.0-alpha1 (2014-12-03)